CA-71 fixed non-updated random_scene() with timeline
Sphere's need to be instantiated with timeline objects, and none of the adds in random_scene did so. Added two lambda functions to create still and randomized timeline and used it. That scene now generates a full scene with 4 large spheres of varying material and a bunch of "bouncing" spheres.

Added some documentation.